chore(mobile): move settings components to UI package
