+++ mobile/lib/widgets/settings/asset_list_settings/asset_list_group_settings.dart
@@ -8,8 +8,7 @@ import 'package:immich_mobile/generated/translations.g.dart';
 import 'package:immich_mobile/providers/app_settings.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/settings.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/timeline.provider.dart';
-import 'package:immich_mobile/widgets/settings/setting_group_title.dart';
-import 'package:immich_mobile/widgets/settings/settings_radio_list_tile.dart';
+import 'package:immich_ui/immich_ui.dart';
 
 class GroupSettings extends HookConsumerWidget {
   const GroupSettings({super.key});
