+++ mobile/lib/widgets/settings/asset_list_settings/asset_list_layout_settings.dart
@@ -6,8 +6,7 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/generated/translations.g.dart';
 import 'package:immich_mobile/providers/app_settings.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/settings.provider.dart';
-import 'package:immich_mobile/widgets/settings/setting_group_title.dart';
-import 'package:immich_mobile/widgets/settings/settings_slider_list_tile.dart';
+import 'package:immich_ui/immich_ui.dart';
 
 class LayoutSettings extends HookConsumerWidget {
   const LayoutSettings({super.key});
