+++ mobile/lib/widgets/settings/preference_settings/preference_setting.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:immich_mobile/widgets/settings/preference_settings/haptic_setting.dart';
 import 'package:immich_mobile/widgets/settings/preference_settings/share_setting.dart';
 import 'package:immich_mobile/widgets/settings/preference_settings/theme_setting.dart';
-import 'package:immich_mobile/widgets/settings/settings_sub_page_scaffold.dart';
+import 'package:immich_ui/immich_ui.dart';
 
 class PreferenceSetting extends StatelessWidget {
   const PreferenceSetting({super.key});
