+++ mobile/lib/widgets/settings/preference_settings/share_setting.dart
@@ -8,8 +8,7 @@ import 'package:immich_mobile/domain/models/settings_key.dart';
 import 'package:immich_mobile/extensions/build_context_extensions.dart';
 import 'package:immich_mobile/generated/translations.g.dart';
 import 'package:immich_mobile/providers/infrastructure/settings.provider.dart';
-import 'package:immich_mobile/widgets/settings/setting_group_title.dart';
-import 'package:immich_mobile/widgets/settings/settings_radio_list_tile.dart';
+import 'package:immich_ui/immich_ui.dart';
 
 class ShareSetting extends HookConsumerWidget {
   const ShareSetting({super.key});
