+++ mobile/packages/ui/lib/immich_ui.dart
@@ -7,6 +7,7 @@ export 'src/components/formatted_text.dart';
 export 'src/components/icon_button.dart';
 export 'src/components/menu_item.dart';
 export 'src/components/password_input.dart';
+export 'src/components/settings/settings.dart';
 export 'src/components/text_button.dart';
 export 'src/components/text_input.dart';
 export 'src/components/url_input.dart';
