+++ mobile/packages/ui/lib/src/components/settings/settings_slider_list_tile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:immich_mobile/extensions/build_context_extensions.dart';
+import 'package:immich_ui/extensions/build_context_extensions.dart';
 
 class SettingsSliderListTile extends StatelessWidget {
   final ValueNotifier<int> valueNotifier;
